Refactored a bit of the main class, added AI class and started working on functionality of it.
